ai-agent/requirements.txt
Werkzeug==2.2.2
Flask==2.2.2
groq==0.18.0
langdetect==1.0.9
iso639-lang==2.6.3
requests==2.32.4